updated enums to use pascal case updated all enums used it many classes created StorageDriveService Created StorageDriveDTO Created StorageDriveRequest Created MetaData model Created HttpClientName enum Replaced all _httpFactory.CreateClient name parameter to use enums converted ToString

+++ src/SimpleStorageSystem.Daemon/Models/Tables/StorageDrive.cs
@@ -4,7 +4,7 @@ namespace SimpleStorageSystem.Daemon.Models.Tables;
 
 public class StorageDrive
 {
-    public int StorageDriveId { get; set; }
+    public long StorageDriveId { get; set; }
     public required string Name { get; set; }
     public MountOption Mount { get; set; }
